+++ packages/autorest.python/ChangeLog.md
@@ -14,6 +14,7 @@
 **Bug Fixes**
 
 - Fix generation failure for `format: password`  #1404
+- Fix `content_type` error when paging with body  #1407
 
 **Other Changes**